# [Feature](new function) Add a uuid-numeric function, returns uuid in largerint type, 20x faster than uuid