--- a/docs/zh-CN/docs/sql-manual/sql-functions/numeric-functions/uuid_numeric.md
+++ b/docs/zh-CN/docs/sql-manual/sql-functions/numeric-functions/uuid_numeric.md
@@ -1,7 +1,7 @@
 ---
 {
     "title": "uuid_numeric",
-    "language": "en"
+    "language": "zh-CN"
 }
 ---