--- a/fe/fe-core/src/main/java/org/apache/doris/nereids/trees/expressions/functions/Nondeterministic.java
+++ b/fe/fe-core/src/main/java/org/apache/doris/nereids/trees/expressions/functions/Nondeterministic.java
@@ -24,7 +24,6 @@ import org.apache.doris.qe.ConnectContext;
  *
  * e.g. 'rand()', 'random()'.
  *
- * note: no 'uuid' function currently.
  */
 public interface Nondeterministic extends ExpressionTrait {
     @Override